+++ Gemfile
@@ -2,7 +2,7 @@
 
 source 'https://rubygems.org'
 
-# Specify your gem's dependencies in railyard.gemspec
+# Specify your gem's dependencies in trokko.gemspec
 gemspec
 
 gem 'rake', '~> 13.0'
