+++ lib/trokko/templates/Dockerfile.erb
