+++ lib/trokko/templates/config/database.yml.erb
