+++ lib/trokko/templates/config/database/mysql.yml.erb
