+++ lib/trokko/templates/config/database/postgresql.yml.erb
