+++ lib/trokko/templates/docker-compose.yml.erb
@@ -16,8 +16,8 @@ volumes:
   database:
     driver: local
     labels:
-      info.snaka.railyard.project: "<%= project_name %>"
+      info.snaka.trokko.project: "<%= project_name %>"
   bundle:
     driver: local
     labels:
-      info.snaka.railyard.project: "<%= project_name %>"
+      info.snaka.trokko.project: "<%= project_name %>"
