+++ lib/trokko/templates/docker_compose/mysql/app_service.erb
