+++ lib/trokko/templates/docker_compose/mysql/db_service.erb
