+++ lib/trokko/templates/docker_compose/postgresql/db_service.erb
