+++ lib/trokko/templates/entrypoint.sh.erb
