+++ lib/trokko/version.rb
@@ -1,5 +1,5 @@
 # frozen_string_literal: true
 
-module Railyard
+module Trokko
   VERSION = '0.1.0'
 end
