+++ sig/trokko.rbs
@@ -1,4 +1,4 @@
-module Railyard
+module Trokko
   VERSION: String
   # See the writing guide of rbs: https://github.com/ruby/rbs#guides
 end
