+++ spec/lib/trokko/scaffolders/config/database_spec.rb
@@ -1,8 +1,8 @@
 # frozen_string_literal: true
 
-require 'railyard/scaffolders/docker_compose'
+require 'trokko/scaffolders/config/database'
 
-RSpec.describe Railyard::Scaffolders::DockerCompose do
+RSpec.describe Trokko::Scaffolders::Config::Database do
   describe '#generate' do
     subject { described_class.new(db:, thor:).generate }
 
@@ -11,13 +11,13 @@ RSpec.describe Railyard::Scaffolders::DockerCompose do
     let(:db) { 'mysql' }
     let(:thor) { thor_dummy }
 
-    it { is_expected.to eq 'dummy/docker-compose.yml' }
-    it { is_expected.to be_a_file_with 'image: mysql:latest' }
+    it { is_expected.to eq 'dummy/config/database.yml' }
+    it { is_expected.to be_a_file_with 'adapter: mysql2' }
 
     context 'when db is PostgreSQL' do
       let(:db) { 'postgresql' }
 
-      it { is_expected.to be_a_file_with 'image: postgres:latest' }
+      it { is_expected.to be_a_file_with 'adapter: postgresql' }
     end
   end
 end
