+++ spec/lib/trokko/scaffolders/docker_compose_spec.rb
@@ -1,8 +1,8 @@
 # frozen_string_literal: true
 
-require 'railyard/scaffolders/config/database'
+require 'trokko/scaffolders/docker_compose'
 
-RSpec.describe Railyard::Scaffolders::Config::Database do
+RSpec.describe Trokko::Scaffolders::DockerCompose do
   describe '#generate' do
     subject { described_class.new(db:, thor:).generate }
 
@@ -11,13 +11,13 @@ RSpec.describe Railyard::Scaffolders::Config::Database do
     let(:db) { 'mysql' }
     let(:thor) { thor_dummy }
 
-    it { is_expected.to eq 'dummy/config/database.yml' }
-    it { is_expected.to be_a_file_with 'adapter: mysql2' }
+    it { is_expected.to eq 'dummy/docker-compose.yml' }
+    it { is_expected.to be_a_file_with 'image: mysql:latest' }
 
     context 'when db is PostgreSQL' do
       let(:db) { 'postgresql' }
 
-      it { is_expected.to be_a_file_with 'adapter: postgresql' }
+      it { is_expected.to be_a_file_with 'image: postgres:latest' }
     end
   end
 end
